+++ src/ScottPlot5/ScottPlot5 Cookbook/Recipes/PlotTypes/Bar.cs
@@ -42,6 +42,33 @@ public class Bar : ICategory
             bars2.Label = "Beta";
 
             myPlot.ShowLegend(Alignment.UpperLeft);
+            myPlot.Axes.Margins(bottom: 0);
+        }
+    }
+
+    public class BarValues : RecipeBase
+    {
+        public override string Name => "Bar with Value Labels";
+        public override string Description => "Set the `Label` property of bars " +
+            "to have text displayed above each bar.";
+
+        [Test]
+        public override void Execute()
+        {
+            double[] values = { 5, 10, 7, 13 };
+            var barPlot = myPlot.Add.Bars(values);
+
+            // define the content of labels
+            foreach (var bar in barPlot.Bars)
+            {
+                bar.Label = bar.Value.ToString();
+            }
+
+            // customize label style
+            barPlot.ValueLabelStyle.Bold = true;
+            barPlot.ValueLabelStyle.FontSize = 18;
+
+            myPlot.Axes.Margins(bottom: 0, top: .2);
         }
     }